+++ src/1c4a/date.ts
@@ -3,7 +3,7 @@ import moment from 'moment-timezone';
 const TIMEZONE = 'Europe/Berlin';
 
 export const formatDate = (date?: Date, timezone = TIMEZONE): string => {
-  date = date || new Date();
+  const dateObj = date || new Date();
 
-  return moment(date).tz(timezone).format('DDMMYYYY-HHmmss');
+  return moment(dateObj).tz(timezone).format('DDMMYYYY-HHmmss');
 };