+++ src/services/DataStore.ts
@@ -1,13 +1,11 @@
 import { Debugger } from 'debug';
+import findCacheDir from 'find-cache-dir';
 import fs from 'fs';
 import { inject, injectable } from 'inversify';
-import { tmpdir } from 'os';
 import { join as joinPath } from 'path';
 // @ts-ignore
-import { version as packageVersion } from '../../package.json';
+import { name as packageName, version as packageVersion } from '../../package.json';
 import { TYPES } from '../di/types';
-
-const BASE_DIR = 'node-internetmarke';
 
 interface CacheFormat<T> {
   date: string;
@@ -29,11 +27,17 @@ export interface IDataStore<T> {
 @injectable()
 export class DataStore<T> implements IDataStore<T> {
   private file: string;
+
   private name: string;
+
   private ttl = 7 * 24 * 3600;
+
   private data: { [id: number]: T };
+
   private lastUpdate: Date | null = null;
+
   private log: Debugger;
+
   private loadData: () => Promise<{ [id: number]: T }>;
 
   constructor(@inject(TYPES.LoggerFactory) getLogger: Function) {
@@ -54,13 +58,10 @@ export class DataStore<T> implements IDataStore<T> {
       this.ttl = ttl;
     }
 
-    const dir = joinPath(tmpdir(), BASE_DIR);
-    try {
-      fs.accessSync(dir);
-    } catch (err) {
-      fs.mkdirSync(dir);
-    }
-    this.file = joinPath(dir, file);
+    const dir = findCacheDir({ name: packageName, create: true });
+    this.log('cache', dir);
+
+    this.file = joinPath(dir!, file);
     fs.closeSync(fs.openSync(this.file, 'a'));
 
     await this.load();
@@ -128,7 +129,7 @@ export class DataStore<T> implements IDataStore<T> {
    */
   protected isValid(): boolean {
     if (!this.lastUpdate || !this.ttl || this.ttl * 1000 < Date.now() - this.lastUpdate.getTime()) {
-      this.log(`store %s is outdated`, this.name);
+      this.log('store %s is outdated', this.name);
 
       return false;
     }